Generates the feeds controller
Uses the controller generator:
 rails g controller feeds index

To create the controller "feeds" with a first action, "index"

The output is:

      create  app/controllers/feeds_controller.rb
       route  get "feeds/index"
      invoke  erb
      create    app/views/feeds
      create    app/views/feeds/index.html.erb
      invoke  rspec
      create    spec/controllers/feeds_controller_spec.rb
      create    spec/views/feeds
      create    spec/views/feeds/index.html.erb_spec.rb
      invoke  helper
      create    app/helpers/feeds_helper.rb
      invoke    rspec
      create      spec/helpers/feeds_helper_spec.rb
      invoke  assets
      invoke    coffee
      create      app/assets/javascripts/feeds.js.coffee
      invoke    scss
      create      app/assets/stylesheets/feeds.css.scss

So not only does it create the controller with the "index" method, it
also creates templates for rendering it.

At this point, the output of `rake cucumber` is: 

Feature: Feeds

  Background:                        # features/feeds.feature:2
    Given the following feeds exist: #
features/step_definitions/steps.rb:1
      | name      | description                    |
      | homicides | zomg killing is pretty bad man |
      | fires     | really dude, arson?            |

  Scenario: View the list of feeds                         #
features/feeds.feature:9
    When I visit the feeds list                            #
features/step_definitions/steps.rb:9
  1 OpenData::Application.routes.draw do
    Then I should see the names for all the existing feeds #
features/step_definitions/steps.rb:13
      expected to find text "homicides" in "Feeds#index Find me in
app/views/feeds/index.html.erb"
(RSpec::Expectations::ExpectationNotMetError)
      ./features/step_definitions/steps.rb:18:in `block (2 levels) in
<top (required)>'
      ./features/step_definitions/steps.rb:14:in `each'
      ./features/step_definitions/steps.rb:14:in `/^I should see the
names for all the existing feeds$/'
      features/feeds.feature:11:in `Then I should see the names for all
the existing feeds'

Failing Scenarios:
cucumber features/feeds.feature:9 # Scenario: View the list of feeds

1 scenario (1 failed)
3 steps (1 failed, 2 passed)
0m1.305s

Which means that the default behavior for an empty action in a rails
controller is to delegate control directly to the view rendering, which
explains why it's expecting to find text in the template already.

More info at
http://guides.rubyonrails.org/action_controller_overview.html