Makes the feeds feature pass
But something is wrong! We skipped testing how the view is generated:
how, for example, is the name getting titleized?

We could test the views
(https://www.relishapp.com/rspec/rspec-rails/v/2-1/docs/view-specs/view-spec),
but, really, having templates with logic makes content negotiation hard:

* What if _every_ format (json, html, xml, hal) should have the feed name titleized?
* What if we change some css classes and the view spec has to be
  maintained to meet them, despite there not being a significant change
  in the data itself?

The solution is to have really dumb templates. But that means having
some object that applies the logic. It could be the model, but, no,
let's not make it the model! (moar fear-mongering links:
http://blog.codeclimate.com/blog/2012/10/17/7-ways-to-decompose-fat-activerecord-models/
)